backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
playwright>=1.48.0
langgraph>=0.2.55
langchain>=0.3.13
langchain-google-genai>=2.0.8
google-generativeai>=0.8.3
python-multipart>=0.0.18
pillow>=11.0.0
python-dotenv>=1.0.1
pydantic>=2.10.5
aiofiles>=24.1.0
structlog>=24.4.0
langchain-core>=0.3.28

